resampledata: Data Sets for Mathematical Statistics with Resampling in R

Package of data sets from "Mathematical Statistics with Resampling in R" (2011) by Laura Chihara and Tim Hesterberg. Note the Bangladesh data is provided solely for illustrative purposes and to enable statistical analysis. Full data are available from the British Geological Survey web site http://www.bgs.ac.uk/arsenic/bphase2/datadownload.htm.

Data sets

Type data(package="resampledata") to view all datasets.

References

Hesterberg, Tim (2014), What Teachers Should Know about the Bootstrap: Resampling in the Undergraduate Statistics Curriculum, http://arxiv.org/abs/1411.5279.

Chihara, Laura M., Hesterberg, Tim C. (2011). Mathematical Statistics with Resampling and R. Hoboken, NJ: John Wiley & Sons.

Examples

library(resampledata)
data(FlightDelays)